Moved file management classes to sub namespace
