+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   The new behavior is, that any entry that appears multiple times in the graph of the response will be the same instance.
 * The `contentful` script used to warm up/clear the cache was not marked as a binary in `composer.json` and thus not published to `vendor/bin`.
 
+### Changed
+* Moved file management classes to a sub-namespace `Contentful\File`.
+
 ## [1.2.0](https://github.com/contentful/contentful.php/tree/1.2.0) (2017-05-16)
 
 ### Added
