+++ tests/Unit/File/FileTest.php
@@ -6,7 +6,7 @@
 
 namespace Contentful\Tests\Unit\Delivery;
 
-use Contentful\File;
+use Contentful\File\File;
 
 class FileTest extends \PHPUnit_Framework_TestCase
 {
